Reformat package date
Changing the date to something that servers and programs can perform math on and therefore easily order.
As it was, users had to scan the list for the latest release- they were ordered alphabetically instead of
by release date. Now the zips will automatically be presented in the order they are made. I also removed
the seconds from the name-- that is unnecessarily precise, no one is building multiple times a minute (yet).

Change-Id: I0e4e92a4d30017659ee865461be1670522c5034c